move registration dashboard logic to utils
