+++ ui/src/app/pages/profile/profile-component.tsx
@@ -15,7 +15,7 @@ import {SpinnerOverlay} from 'app/components/spinners';
 import {AoU} from 'app/components/text-wrappers';
 import {withProfileErrorModal, WithProfileErrorModalProps} from 'app/components/with-error-modal';
 import {WithSpinnerOverlayProps} from 'app/components/with-spinner-overlay';
-import {getRegistrationTasksMap} from 'app/pages/homepage/registration-dashboard';
+import {getRegistrationTasksMap} from 'app/utils/access-utils';
 import {AccountCreationOptions} from 'app/pages/login/account-creation/account-creation-options';
 import {DataAccessPanel} from 'app/pages/profile/data-access-panel';
 import {DemographicSurvey} from 'app/pages/profile/demographic-survey';
